refactor(fe): move dataIfError.ts file to _libs folder
