+++ apps/frontend/app/(client)/(code-editor)/problem/[problemId]/submission/_components/SubmissionDetail.tsx
@@ -11,7 +11,7 @@ import {
 import { dateFormatter, fetcherWithAuth, getResultColor } from '@/lib/utils'
 import type { SubmissionDetail } from '@/types/type'
 import { revalidateTag } from 'next/cache'
-import dataIfError from './dataIfError'
+import dataIfError from '../_libs/dataIfError'
 
 interface Props {
   problemId: number
